DIV-6317: clean up code
