+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/bulk/print/dataextractor/CoECoverLetterDataExtractor.java
@@ -4,9 +4,10 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.BulkCaseConstants;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants;
+import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.Gender;
+import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.Relation;
 
 import java.util.Map;
-import java.util.Optional;
 
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.YES_VALUE;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper.ExtractorHelper.getMandatoryStringValue;
@@ -16,18 +17,26 @@ public class CoECoverLetterDataExtractor {
 
     @NoArgsConstructor(access = AccessLevel.PRIVATE)
     public static class CaseDataKeys {
+        public static final String COURT_NAME = BulkCaseConstants.COURT_NAME_CCD_FIELD;
+        public static final String HEARING_DATE = OrchestrationConstants.DATE_OF_HEARING_CCD_FIELD;
+        public static final String HEARING_DATE_TIME = OrchestrationConstants.DATETIME_OF_HEARING_CCD_FIELD;
+        public static final String PETITIONER_GENDER = OrchestrationConstants.D_8_INFERRED_PETITIONER_GENDER;
         public static final String COST_CLAIMED = OrchestrationConstants.DIVORCE_COSTS_CLAIM_GRANTED_CCD_FIELD;
         public static final String COURT_ID = BulkCaseConstants.COURT_NAME_CCD_FIELD;
     }
 
-    public static boolean isCostsClaimGranted(Map<String, Object> caseData) {
-        return Optional.ofNullable(caseData.get(CaseDataKeys.COST_CLAIMED))
-            .map(String.class::cast)
-            .map(YES_VALUE::equalsIgnoreCase)
-            .orElse(false);
-    }
-
     public static String getCourtId(Map<String, Object> caseData) {
         return getMandatoryStringValue(caseData, CaseDataKeys.COURT_ID);
+    }
+
+    public static String getHusbandOrWife(Map<String, Object> caseData) {
+        Gender petitionerInferredGender = Gender.from(getMandatoryStringValue(caseData, CaseDataKeys.PETITIONER_GENDER));
+        Relation petitionerRelationshipToRespondent = Relation.from(petitionerInferredGender);
+
+        return petitionerRelationshipToRespondent.getValue();
+    }
+
+    public static boolean isCostsClaimGranted(Map<String, Object> caseData) {
+        return YES_VALUE.equalsIgnoreCase((String) caseData.get(CaseDataKeys.COST_CLAIMED));
     }
 }
