+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/bulk/print/dataextractor/SolicitorDataExtractor.java
@@ -1,13 +1,22 @@
 package uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor;
 
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants;
+
 import java.util.Map;
 
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper.ExtractorHelper.getMandatoryStringValue;
 
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class SolicitorDataExtractor {
 
-    public static String getSolicitorReference(Map<String, Object> caseData) {
-        return getMandatoryStringValue(caseData, CoELetterDataExtractor.CaseDataKeys.SOLICITOR_REFERENCE);
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    public static class CaseDataKeys {
+        public static final String SOLICITOR_REFERENCE = OrchestrationConstants.D8_RESPONDENT_SOLICITOR_REFERENCE;
     }
 
+    public static String getSolicitorReference(Map<String, Object> caseData) {
+        return getMandatoryStringValue(caseData, CaseDataKeys.SOLICITOR_REFERENCE);
+    }
 }
